requirements.txt
crewai
CrewaAI[tools]
python-dotenv
duckduckgosearchrun
streamlit